requirements.txt
alive_progress
openai==0.28
pandas
openai_multi_client
python-dotenv
scikit-learn
